Changed folder structure for imageCarousel and logoData
- Instances of these two types of data will now be in their own file for their specific page so the logoData/imageCarouselData files aren't cluttered with instances meant for different pages
