+++ website/src/lib/data/logoData/personalSiteLogoData.ts
@@ -1,0 +1,15 @@
+import { LogoInterface } from "@/interface/logoInterface";
+import GitHubIcon from "@/public/github-mark-white.svg";
+
+export const personalSiteLogoData: LogoInterface[] = [
+  {
+    id: "personalSiteLogo-1",
+    link: "https://github.com/alexrojas431/alexrojas431.github.io",
+    image: {
+      id: "personalSiteLogo-1-img",
+      src: GitHubIcon,
+      alt: "GitHub Icon, click to go to the projects GitHub Page",
+    },
+    description: "Project code on GitHub",
+  },
+];
