+++ website/src/lib/data/logoData/reactJam23_24WinterLogoData.ts
@@ -1,0 +1,15 @@
+import { LogoInterface } from "@/interface/logoInterface";
+import GitHubIcon from "@/public/github-mark-white.svg";
+
+export const reactJam23_24WinterLogoData: LogoInterface[] = [
+  {
+    id: "reactJam23_24Winter-Logo-1",
+    link: "https://github.com/alexrojas431/CyberpunkTycoon",
+    image: {
+      id: "reactJam23_24Winter-Logo_1-img",
+      src: GitHubIcon,
+      alt: "GitHub Icon, click to go to the projects GitHub Page",
+    },
+    description: "Project code on GitHub",
+  },
+];
